test functions for entropic
